refactor: UI  역할별 파일 분리 및 함수 분리
