+++ src/js/screen/screenManager.js
@@ -1,0 +1,11 @@
+import ModalManualScreen from './modalManualScreen.js';
+import SearchVideoScreen from './searchVideoScreen.js';
+import VideoManualScreen from './videoManualScreen.js';
+
+export default class ScreenManager {
+  constructor() {
+    new ModalManualScreen();
+    new SearchVideoScreen();
+    new VideoManualScreen();
+  }
+}
