+++ src/js/util/constants.js
@@ -1,9 +1,7 @@
-export const SEARCH_NOT_WORKING_ERROR_MESSAGE = '유튜브 검색 기능이 정상 작동되지 않았습니다.';
-
 export const MAX_SAVED_VIDEOS_LENGTH = 100;
 
-export const DELAY_MILISECOND_TIME = 500;
-
 export const VIDEO_COUNT = 10;
+
+export const DELAY_MILISECOND_TIME = 500;
 
 export const SERVER_ERROR_STATUS = '403';
